refactor: Update LlmType enum values and improve error message for OneUptime LLM server not reachable
