+++ Copilot/Types/LlmType.ts
@@ -1,5 +1,5 @@
 enum LlmType {
-  LLM = "OneUptime LLM Server", // OneUptime custom LLM Server
+  ONEUPTIME_LLM = "OneUptime LLM Server", // OneUptime custom LLM Server
   OpenAI = "OpenAI",
 }
