+++ Copilot/Utils/Init.ts
@@ -18,7 +18,7 @@ import logger from "Common/Server/Utils/Logger";
 
 export default class InitUtil {
   public static async init(): Promise<CodeRepositoryResult> {
-    if (GetLlmType() === LlmType.LLM) {
+    if (GetLlmType() === LlmType.ONEUPTIME_LLM) {
       const llmServerUrl: URL = GetLlmServerUrl();
       // check status of ll, server
       const result: HTTPErrorResponse | HTTPResponse<JSONObject> =
@@ -26,7 +26,7 @@ export default class InitUtil {
 
       if (result instanceof HTTPErrorResponse) {
         throw new BadDataException(
-          "Llama server is not reachable. Please check the server URL in the environment variables.",
+          "OneUptime LLM server is not reachable. Please check the server URL in the environment variables.",
         );
       }
     }
