+++ Copilot/Utils/Process.ts
@@ -1,0 +1,5 @@
+export default class ProcessUtil {
+  public static haltProcessWithSuccess(): void {
+    process.exit(0);
+  }
+}
